Внесены правки, вместо целого числа можно ставить дробные

+++ ideal_weight.rb
@@ -2,12 +2,6 @@
 # выводит результат пользователю на экран с обращением по имени. Если идеальный вес получается отрицательным,
 # то выводится строка "Ваш вес уже оптимальный"
 
-
-# Запрашиваем имя рост
-# Считаем вес по формуле
-# Выводим результат
-
-
 def ask_name
   puts "please, write your name: "
   name = gets.strip
@@ -16,7 +10,7 @@ end
 
 def ask_height
   puts "please, write your height: "
-  height = gets.to_i
+  height = gets.to_f
   return height
 end
 
@@ -25,11 +19,9 @@ def calculate_weight(height)
   return ideal
 end
 
-
 user_name = ask_name
 user_height = ask_height
 user_ideal_weight = calculate_weight(user_height)
-
 
 if user_ideal_weight > 0
   puts "#{user_name} your ideal weight is #{user_ideal_weight}"
